% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\docType{data}
\name{path_params}
\alias{path_params}
\title{list of allowed path parameter names}
\format{
An object of class \code{character} of length 7.
}
\usage{
path_params
}
\description{
list of allowed path parameter names
}
\keyword{datasets}
