% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\docType{data}
\name{query_params}
\alias{query_params}
\title{list of allowed query parameter names}
\format{
An object of class \code{character} of length 2.
}
\usage{
query_params
}
\description{
list of allowed query parameter names
}
\keyword{datasets}
