% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{wl_parse}
\alias{wl_parse}
\title{parse output from rweatherlink objects}
\usage{
wl_parse(x, ...)
}
\arguments{
\item{x}{weatherlink object}

\item{...}{other arguments to methods (currently unused)}
}
\description{
parse output from rweatherlink objects
}
\examples{
\dontrun{
(stations <- wl_stations() \%>\% wl_parse())
wl_nodes() \%>\% wl_parse()
wl_sensors() \%>\% wl_parse()
wl_sensor_activity() \%>\% wl_parse()
wl_current(station_id = stations$station_id[[1]]) \%>\% wl_parse()
wl_historic(station_id = stations$station_id[[1]],
  start = as.integer(as.POSIXct("2020-06-16 13:59:39 CET")),
  end = as.integer(as.POSIXct("2020-06-16 20:59:39 CET"))
 ) \%>\% wl_parse()
}
}
