% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{wl_request}
\alias{wl_request}
\alias{wl_params}
\title{General request function for weatherlink api}
\usage{
wl_request(
  endpoint = NULL,
  path = NULL,
  query = NULL,
  api_key = Sys.getenv("WL_API_KEY"),
  api_secret = Sys.getenv("WL_API_SECRET"),
  ignore_ratelimit = FALSE,
  verbose = FALSE
)

wl_params(
  path = NULL,
  query = NULL,
  api_key = Sys.getenv("WL_API_KEY"),
  api_secret = Sys.getenv("WL_API_SECRET")
)
}
\arguments{
\item{endpoint}{api endpoint to query}

\item{path}{named list; path parameter to pass}

\item{query}{named list of length 2; Only "start-timestamp", "end_timestamp"
are allowed}

\item{api_key}{API key. Read by default from env variable \code{WL_API_KEY}}

\item{api_secret}{API secret. Read by default from env variable
\code{WL_API_SECRET}}

\item{ignore_ratelimit}{ignore wait time to not hit rate limit of 10/second?}

\item{verbose}{logical, should the request be printed?}
}
\description{
authentication is done via hmac, see \code{\link[=wl_params]{wl_params()}}.
}
\examples{
\dontrun{
stations <- wl_request(endpoint = "stations")
wl_request(endpoint = "stations",
  path = list(`station-ids` = stations$stations[[1]]$station_id))
wl_request(endpoint = "stations",
  path = list(`station-ids` = c(stations$stations[[1]]$station_id,
    stations$stations[[2]]$station_id)))
wl_request(endpoint = "current",
  path = list(`station-id` = stations$stations[[1]]$station_id))
wl_request(endpoint = "historic",
  path = list(`station-id` = stations$stations[[1]]$station_id),
  query = list(
    `start-timestamp` = as.integer(as.POSIXct("2020-06-16 13:59:39 CET")),
    `end-timestamp` = as.integer(as.POSIXct("2020-06-16 20:59:39 CET"))))
}
\dontrun{
wl_params()
wl_params(path = list(`station-ids` = 123456))
}
}
