% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wl_stations}
\alias{wl_stations}
\alias{wl_nodes}
\alias{wl_sensors}
\alias{wl_sensor_activity}
\alias{wl_sensor_catalog}
\alias{wl_current}
\alias{wl_historic}
\title{Query /stations & stations/{station-ids} endpoint}
\usage{
wl_stations(station_ids = NULL, ...)

wl_nodes(node_ids = NULL, ...)

wl_sensors(sensor_ids = NULL, ...)

wl_sensor_activity(sensor_ids = NULL, ...)

wl_sensor_catalog(...)

wl_current(station_id, ...)

wl_historic(station_id, start, end, ...)
}
\arguments{
\item{station_ids}{vector of stations ids to query. If NULL (default) all
will be returned.}

\item{...}{other arguments passed to \code{\link[=wl_request]{wl_request()}}}

\item{node_ids}{vector of stations ids to query. If NULL (default) all will
be returned.}

\item{sensor_ids}{vector of stations ids to query. If NULL (default) all
will be returned.}

\item{station_id}{station_id for which to query data.}

\item{start}{integer; Unix timestamp. Current timestamp can be generated
with as.integer(Sys.time())}

\item{end}{integer; Unix timestamp. Current timestamp can be generated
with as.integer(Sys.time())}
}
\description{
Query /stations & stations/{station-ids} endpoint

Query /nodes & nodes/{nodes-ids} endpoint

Query /sensors & sensors/{sensor-ids} endpoint

Query /sensor-activity & sensor-activity/{sensor-ids} endpoint

Query /sensor-catalog

Query /current/{station-id} endpoint

Query /historic/{station-id} endpoint
}
\examples{
\dontrun{
(stations <- wl_stations())
wl_stations(station_ids = stations$stations[[1]]$station_id)
wl_stations(station_ids = c(stations$stations[[1]]$station_id,
    stations$stations[[2]]$station_id))
}
\dontrun{
wl_nodes()
}
\dontrun{
(sensors <- wl_sensors())
wl_sensors(sensor_ids = sensors$sensors[[1]]$lsid)
wl_sensors(sensor_ids = c(sensors$sensors[[1]]$lsid,
  sensors$sensors[[2]]$lsid))
}
\dontrun{
wl_sensor_activity()
wl_sensor_activity(sensor_ids = sensors$sensors[[1]]$lsid)
wl_sensor_activity(sensor_ids = c(sensors$sensors[[1]]$lsid,
  sensors$sensors[[2]]$lsid))
}
\dontrun{
wl_sensor_catalog()
}
\dontrun{
(stations <- wl_stations())
wl_current(station_id = stations$stations[[1]]$station_id)
}
\dontrun{
(stations <- wl_stations())
wl_historic(station_id = stations$stations[[1]]$station_id,
  start = as.integer(as.POSIXct("2020-06-16 13:59:39 CET")),
  end = as.integer(as.POSIXct("2020-06-16 20:59:39 CET"))
 )
}
}
\references{
https://weatherlink.github.io/v2-api/api-reference
}
